refactor: use decimals where necessary instead of integers

+++ api/gql/deal/inputs.py
@@ -1,5 +1,6 @@
 from datetime import date, datetime
 from typing import Annotated
+from decimal import Decimal
 
 import strawberry
 
@@ -16,8 +17,8 @@ class CreateDealInput:
     seller_id: int | None = None
     buyer_id: int | None = None
 
-    total_amount: int
-    remaining_amount_due: int
+    total_amount: Decimal
+    remaining_amount_due: Decimal
     type: str
 
     installments_total_amount: int = 0
