+++ application/deal/dtos.py
@@ -1,5 +1,6 @@
 from dataclasses import dataclass
 from datetime import date, datetime
+from decimal import Decimal
 
 from core.errors import InvalidError
 
@@ -18,8 +19,8 @@ class CreateDealUsecaseDto:
     seller_id: int | None
     buyer_id: int | None
 
-    total_amount: int
-    remaining_amount_due: int
+    total_amount: Decimal
+    remaining_amount_due: Decimal
     type: str
 
     installments_total_amount: int = 0
